requirements.txt
annotated-types==0.7.0
confluent-kafka==2.8.0
dnspython==2.7.0
email_validator==2.2.0
Faker==35.2.0
idna==3.10
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
typing_extensions==4.12.2
